requirements.txt
attr==0.3.2
colorama==0.4.6
ConfigParser==7.1.0
cryptography==44.0.0
Cython==3.0.11
dl==0.1.0
docutils==0.21.2
filelock==3.16.1
Flask==3.1.0
gradio==5.9.1
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.5
jnius==1.1.0
keyring==25.5.0
numpy==2.2.0
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
ordereddict==1.1
Pillow==11.0.0
protobuf==5.29.1
pyOpenSSL==24.3.0
redis==5.2.1
Sphinx==8.1.3
thread==2.0.5
ultralytics==8.3.50
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
